api/requirements.txt
apscheduler==3.10.4
scipy==1.13.1
scikit-learn==1.6.1
fastapi==0.108.0
fastapi[all]==0.108.0
fastapi-mail==1.4.1
flow-py-sdk==2.0.1
fuzzywuzzy==0.18.0
geonamescache==2.0.0
graphene==3.3
motor==3.3.2
numpy==2.0.2
PyJWT==2.8.0
pymongo==4.8.0
python-Levenshtein==0.27.1
requests==2.31.0
schedule==1.1.0
starlette-graphene3==0.6.0
uvicorn==0.15.0